refining the way that _taggers and _lingpipe align labels from something like hyperlink_labels into Labels on Tokens.
The result is a richer config file that still must be set to use LINE-type offsets
for aligning, because hyperlink_labels is only reliable with the LINE-type parser.

In principle, the new BYTE-type aligner added in this commit also works, but
we do not have any byte-type offsets for testing.