adding --detialed flag for
    kba.pipeline.load --count --detail

which does full task scan (very slow).  If you leave off the --count,
it doesn't count partials nor indicate those in an inconsistent state
by comparing counts, but it is much faster.

This also fixes an issue in which data['end_count'] could be set to
None instead of 0.  I think this only happens when there is strong
competition for tasks, and have not figured out a good test for it.