Add a backoff time for task contention.  Specification was for exponential backoff from 2 to a max of 128 seconds.
Also this serves as an example of how to test functions that are
potentially complex or expensive. We don't really want to have a unit
test wait around for over 128 seconds to see if the max waiting time is
observed.  So we subclass the class and stub out all the functionality
we don't care to test and add a fake sleep function that logs the time
it would have slept and then returns immediately.