Don't keep retrying on OSError
If the pipeline is out of file handles we need to fail hard.

Before this we retried on a task for the configured number of times then
marked it as permanently failed and moved onto the next task.  Each new
task would eventually fail because we never did anything to
fix the underlying problem.