Represent new information in labels yaml file
1) Currently the yaml file which configures the labels and mentions has no
way of structuring the cyber mentions (i.e. indicators) – there is just
one list of all mentions that should be identified and labelled.

This task involves devising a structure for the cyber mentions in the
yaml file and fixing the code which reads it.

Eventually the information in this extra structure will be carried
forward into the streamitems. For now, this information can be
discarded.

2) Modify the labeling yaml file so that we can designate selected file(s)
as one the profile we are using as the labeling target.

closes DIFFEO-39