Make the from_spinn3r_feed stage take the prefetch dictionary as a param.
This should avoid problems running tests concurrently against the shared
prefetch dictionary.  There is no way in the current pipeline to pass in
the dictionary, but there was no way before either; you can set the
prefetch dictionary on the reader stage of a constructed-but-not-executed
pipeline (see the test for an example).