Add a default for tmp_dir_path and third_dir_path, change their handling.
Instead of modifying the global configuration at startup to push these
into every stage, add a PipelineFactory.create() method that pushes them
into the constructor parameters at the time the stage is created.  And
instead of modifying tmp_dir_path in configuration to be a subdirectory
of what the user specified, add an attribute to PipelineFactory that
gets merged in at creation time.

The upshot of this is that a single process with multiple threads can
safely create multiple pipelines from a single configuration/factory.