Thou shalt not slice UTF-8 encoded byte strings.
If the 60th byte of an extracted title in the middle of a multi-byte
encoding of a codepoint, then the resulting trimmed string would fail
to decode.

The fix is to make sure the string is Unicode first, and then slice it,
which will always produce a valid trimmed Unicode string.