Add a kvlayer by_time index.
This is a kvlayer stream item index where keys are (epoch_ticks:int,
doc_id:str) tuples and values are empty, the opposite of the keys for
the main stream item table.

The attentive will note this is the key order the stream items table
used to have.  It is harder to distribute in Bigtables-style databases
like Accumulo, where you can pre-split the table on the MD5-valued
doc_id, but easier to scan documents by time.