Add force_clean_html stage.
This transform will produce clean_html on a stream item that only has
clean_visible filled in. This is useful as a transform that runs late
in a pipeline to ensure that clean_html is truly set.

If it does end up generating clean_html, then it will also re-generate
clean_visible to maintain the invariants between the two.