Move test_pipeline_coordinate out of the way.
It depends on coordinate.workers.MultiWorker, which got removed.
It has only one test, but it unconditionally skips itself, so not trying
to run it doesn't cost anything.  It's much more of a system test than
a unit test; it fires up a coordinate worker (hard-coded as MultiWorker
and specifically invoked via multiprocessing) and only looks at coordinate
external calls to try to see what's going on.

That having been said, as system tests go, it doesn't look bad and at
some point this should get run again.